Create a project directory based on a template.

Create a project directory and associated files. Is based on using RStudio and the devtools package development way of working with files and code. So commands/keybindings like devtools::load_all() (Ctrl-Shift-L) and devtools::document() (Ctrl-Shift-D) work to update your code and allow an continuous and integrated approach to analyzing your data and writing up your research product.

prodigen(type, name = NULL, path = getwd(), git.init = FALSE)

Arguments

type
type of project to create, based on the templates found in prodigenr::template_list() function

name
name of the new project directory

path
location of where the new project will be created

git.init
Logical, whether to initialize the new project with git

Value

Creates a project directory with files and subdirectories

Examples

# NOT RUN {
\donttest{
# Get a list of possible project templates
template_list
# Create a project. Best done in a fresh R console.
prodigen('poster', 'poster', 'dev/', TRUE, FALSE)
prodigen('poster', 'poster', 'path/to/dir/', FALSE)
prodigen('poster', 'posterName', git.init = FALSE)
prodigen('abstract', 'name', './', FALSE)
prodigen('manuscript', path = './', git.init = TRUE)
}
# }